requirements1.txt
appdirs==1.4.3
certifi==2017.11.5
chardet==3.0.4
ConcurrentLogHandler==0.9.1
DBUtils==1.2
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.20.1
greenlet==0.4.12
idna==2.6
jpush==3.2.8
MySQL-python==1.2.5
packaging==16.8
pyparsing==2.2.0
requests==2.18.4
six==1.10.0
urllib3==1.22
websocket-client==0.40.0
